Doors and Key Fixes
This fixes the Door animation so that when you enter a door and exit around the same general area, the door you entered from closes automatically.
Also added some fixes for the Level Editor where after leaving the playtest mode, the number of keys you have resets to 0 and the locked doors re-lock.