Updated Lost Levels Castle WATER Tileset (i forgor) (#388)
* Updated Castle WATER Tileset (i forgor)

Yeah I realized late at night that I forgot to update the Underwater Castle tileset.
This updates the Castle tileset, but underwater.
Basically the same as the last one with a couple of extra changes:
- Linked Deco Tiles for ANN to LL since there's no mushrooms
- Replaced ROM Extractor check for ANN tileset
- Yeah that's about it

Side note only partially related to this, someone should make a Settings button that JUST regenerates the Base Assets without having to re-install the ROM.

* Fucked up a pixel on the regular Brick Block