Added Coin Sparkle Effects (#350)
* Just the Coin Sparkle stuff

FYI, if a property in the json is formatted like `exa.exb` then it reads `exa` as another object and `exb` as a property in `exa` assuming the root `property_node` contains this object.
Also primarily for Resource Packs to change how the particles work and should explain how I got there with my personal SMAS skin.

* Added Coin Sparkle Texture to Scene

Of course I broke it in some form.